# Improve perf for dynamic index / fieldname (Restore the generated function perf)