Fix a regression introduced in #28.

We need to do the field => index lookup at compiletime, but we cannot guarantee we'll get a compiler-constant fieldname.

So we precompute the entire lookup table (as a NamedTuple), then do the lookup in there. If the name is a compiler-constant, great, the lookup compiles away. But even if not, the lookup is very cheap.

Before:
```julia
julia> @btime getindex($foo, $2)
  209.277 ns (2 allocations: 48 bytes)
Blob{Float32}(Ptr{Nothing} @0x000000014b1feed0, 8, 12)

julia> @btime getproperty($foo, $(:x))
  1.833 μs (5 allocations: 176 bytes)
Blob{Int64}(Ptr{Nothing} @0x000000014b1feed0, 0, 12)

julia> @btime getproperty($foo, $(:y))
  1.275 μs (4 allocations: 128 bytes)
Blob{Float32}(Ptr{Nothing} @0x000000014b1feed0, 8, 12)
```

After:
```julia
julia> @btime getindex($foo, $2)
  85.196 ns (2 allocations: 48 bytes)
Blob{Float32}(Ptr{Nothing} @0x000000014b1feed0, 8, 12)

julia> @btime getproperty($foo, $(:x))
  99.912 ns (2 allocations: 48 bytes)
Blob{Int64}(Ptr{Nothing} @0x000000014b1feed0, 0, 12)

julia> @btime getproperty($foo, $(:y))
  107.028 ns (2 allocations: 48 bytes)
Blob{Float32}(Ptr{Nothing} @0x000000014b1feed0, 8, 12)
```

Probably the much more likely/relevant case would be a field-access on a type-unstable blobs value. For example `x.y` where `x` does not have a known type. This was very expensive before this PR:
```julia
julia> @btime ((a)->a[1].y)($(Any[foo]))
  1.458 μs (8 allocations: 352 bytes)
Blob{Float32}(Ptr{Nothing} @0x000000016fe73d90, 8, 12)
```
and now it's reasonable again:
```julia
julia> @btime ((a)->a[1].y)($(Any[foo]))
  115.105 ns (2 allocations: 48 bytes)
Blob{Float32}(Ptr{Nothing} @0x000000016fe73d90, 8, 12)
```